separates models.py into Article.py, Author.py, ArticleHit.py
