+++ articleapp/models/Author.py
@@ -1,0 +1,5 @@
+from django.db import models
+
+
+class Author(models.Model):
+    name = models.TextField()
